+++ core/src/main/java/arez/annotations/Priority.java
@@ -15,9 +15,14 @@ package arez.annotations;
 public enum Priority
 {
   /**
-   * Highest user controllable priority.
+   * Highest priority.
    * This priority should be used when the reaction will dispose other reactive elements (and thus they
-   * need not be scheduled) or if the reaction will trigger many downstream reactions.
+   * need not be scheduled).
+   */
+  HIGHEST,
+  /**
+   * High priority.
+   * This priority should be used when the reaction will trigger many downstream reactions.
    */
   HIGH,
   /**